backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
openai==1.51.0
yfinance==0.2.43
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
